change model

+++ pisces/model.py
@@ -1,4 +1,4 @@
-__COMPLETER_all_completions# -*- coding:utf-8 -*-
+# -*- coding:utf-8 -*-
 
 #import sqlalchemy
 from sqlalchemy.ext.declarative import declarative_base
@@ -52,9 +52,9 @@ class ModelMixin(object):
        the child class must have the expire prop
        the child class must have the uid prop
        the child class must have the model prop
+       the child class must have the cls_orm prop
        default key is uid, default uid is id
     """
-
     def __init__(self):
         self.orm = None
 
@@ -63,7 +63,14 @@ class ModelMixin(object):
             return None
 
         return getattr(self.orm, name)
- 
+
+    @property
+    def uid(self):
+        if hasattr(self, 'orm'):
+            return self.orm.id
+        else:
+            return None
+
     @classmethod
     def _create_from_orm(cls, md):
         """md is sqlalchemy obj"""
@@ -71,20 +78,17 @@ class ModelMixin(object):
         obj.orm = md
         return obj
 
-    @property
-    def uid(self):
-        if hasattr(self, 'orm')
-            return self.orm.id
-        else:
-            return None
-
-    def get_obj_dic(self):
+    def _get_prop_dic(self):
         dic = {}
-        cs = type(self)
-        for n, v in inspect.getmembers(cs):
-            if isinstance(v, Column) and hasattr(self, n):
-                dic[v] = getattr(self, n)
+        for n in self.orm.__dict__:
+            if n in self.cls_orm.__dict__:
+                dic[n] = getattr(self, n)
         return dic
+
+    def _set_prop_dic(self, dic):
+        for n in dic:
+            if n in self.orm.__dict__ and n in self.cls_orm.__dict__:
+                setattr(self.orm, n, dic[n])
 
     @classmethod
     def is_cache_enable(cls):
@@ -92,8 +96,7 @@ class ModelMixin(object):
     
     @classmethod
     def get_cache_key(cls, k):
-        return cache_key_prefix_by_db() + cls.__tablename__ \
-          + '_' + str(k)
+        return cache_key_prefix_by_db() + cls.__name__ + '_' + str(k)
 
     @classmethod
     def get_cache(cls, k):
@@ -119,7 +122,7 @@ class ModelMixin(object):
 
     @classmethod
     def count(cls):
-        return db().query(cls).count()
+        return db().query(cls.cls_orm).count()
 
     @classmethod
     def get(cls, uid):
@@ -130,52 +133,51 @@ class ModelMixin(object):
         if data:
             md = pickle.loads(data)
         else:
-            md = db().query(cls).get(uid)
+            md = db().query(cls.cls_orm).get(uid)
+            if not md:
+                return None
             cls.add_cache(ck, pickle.dumps(md))
         return cls._create_from_orm(md)
 
-    @classmethod
-    def add(cls, obj):
-        cls.add_by_uid(v.uid, v)
+    def update(self):
+        md = db().query(self.cls_orm).get(self.uid)
+        self._create_from_orm(md)._set_prop_dic(self._get_prop_dic())
+        db().flush()
+
+        ck = self.get_cache_key(self.uid)
+        self.set_cache(ck, pickle.dumps(self.orm))
 
     @classmethod
-    def add_by_uid(cls, uid, obj):
+    def add(cls, obj):
+        if not obj:
+            return
+
         db().add(obj.orm)
         db().flush()
 
-        ck = cls.get_cache_key(uid)
-        cls.add_cache(ck, pickle.dumps(v.orm))
+        ck = cls.get_cache_key(obj.uid)
+        cls.add_cache(ck, pickle.dumps(obj.orm))
 
     @classmethod
     def delete(cls, obj):
-        cls.del_by_uid_obj(obj.uid, obj)
-
-    @classmethod
-    def del_by_uid(cls, uid):
-        md = db().query(cls).get(uid)
-
-        db().delete(md)
-        db().flush()
-
-        ck = cls.get_cache_key(uid)
-        cls.del_cache(ck)
-
-    @classmethod
-    def del_by_uid_obj(cls, uid, obj):
+        if not obj:
+            return
+        
         db().delete(obj.orm)
         db().flush()
 
-        ck = cls.get_cache_key(uid)
+        ck = cls.get_cache_key(obj.uid)
         cls.del_cache(ck)
 
     @classmethod
-    def update_by_uid_value(cls, uid, obj):
-        dic = obj.get_prop_dic()
-        db().query(cls).get(uid).update(dic)
-        db().flush()
+    def del_by_uid(cls, uid):
+        if not uid:
+            return
+        
+        md = db().query(cls.cls_orm).get(uid)
+        if md:
+            db().delete(md)
+            db().flush()
 
         ck = cls.get_cache_key(uid)
-        cls.set_cache(ck, pickle.dumps(obj.orm))
-
-    def update(self):
-        self.update_by_key_value(self.uid, self)
+        cls.del_cache(ck)
